+++ src/main/java/com/aleksey/castlegates/config/ConfigManager.java
@@ -3,7 +3,7 @@
  *
  */
 
-package com.aleksey.castlegates.manager;
+package com.aleksey.castlegates.config;
 
 import java.util.Arrays;
 import java.util.HashSet;
