+++ src/main/java/com/aleksey/castlegates/plugins/jukealert/IJukeAlertManager.java
@@ -1,4 +1,4 @@
-package com.aleksey.castlegates.jukealert;
+package com.aleksey.castlegates.plugins.jukealert;
 
 import org.bukkit.Location;
