+++ src/main/java/com/aleksey/castlegates/plugins/orebfuscator/OrebfuscatorManager.java
@@ -1,4 +1,4 @@
-package com.aleksey.castlegates.orebfuscator;
+package com.aleksey.castlegates.plugins.orebfuscator;
 
 import java.util.List;
